Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main() {
-    float  a,b,c;
-    scanf("%f" "%f" "%f",&a &b &c);
+    int  a,b,c;
+    scanf("%d" "%d" "%d",&a &b &c);
     float average= (a+b+c)/3;
     printf("Average: %.2f");
     return 0;
